internal: Move internal folder to library
The internal folder restricts imports to packages in its parent
directory only. Since internal/client is used both by swupd-extract and
swupd-inspector, dep ensure complains that the use is out of scope.

This PR changes internal to library so that the code can be shared.
