+++ swupd-inspector/diff.go
@@ -6,7 +6,7 @@ import (
 	"path/filepath"
 	"sort"
 
-	"github.com/clearlinux/mixer-tools/internal/client"
+	"github.com/clearlinux/mixer-tools/library/client"
 	"github.com/clearlinux/mixer-tools/swupd"
 )
