+++ swupd-inspector/get.go
@@ -8,7 +8,7 @@ import (
 	"path/filepath"
 	"strings"
 
-	"github.com/clearlinux/mixer-tools/internal/client"
+	"github.com/clearlinux/mixer-tools/library/client"
 	"github.com/clearlinux/mixer-tools/swupd"
 )
