+++ swupd-inspector/log.go
@@ -5,7 +5,7 @@ import (
 	"log"
 	"path/filepath"
 
-	"github.com/clearlinux/mixer-tools/internal/client"
+	"github.com/clearlinux/mixer-tools/library/client"
 	"github.com/clearlinux/mixer-tools/swupd"
 )
